requirements.txt
# Core ML
torch>=2.0
torchvision>=0.15

# Data handling
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2

# Image processing
Pillow>=9.5
opencv-python>=4.8

# Training utilities
tqdm>=4.65
matplotlib>=3.7

# Experiment tracking
tensorboard>=2.13

# Demo
streamlit>=1.25
